+++ backend/main.go
@@ -14,6 +14,12 @@ import (
 func main() {
 	log.Logger = log.Output(zerolog.ConsoleWriter{Out: os.Stderr})
 
+	level, err := zerolog.ParseLevel(utils.GetEnv("LOG_LEVEL", "info"))
+	if err != nil {
+		level = zerolog.InfoLevel
+	}
+	zerolog.SetGlobalLevel(level)
+
 	a := &App{}
 
 	httpPort := utils.GetEnv("HTTP_PORT", "8080")