+++ backend/model.go
@@ -25,11 +25,5 @@ type Votes struct {
 
 type Offer struct {
 	DealerName string `json:"dealer_name"`
-	Url string `json:"url"`
+	Url        string `json:"url"`
 } // @name Offer
-
-type TopicDetails struct {
-	Topic       Topic  `json:"topic"`
-	Description string `json:"description"`
-	FirstPost   string `json:"first_post"`
-} // @name TopicDetails