+++ vite.config.mjs
@@ -25,7 +25,7 @@ export default defineConfig({
   server: {
     port: 3000,
     proxy: {
-      "/api": "http://localhost:8080",
+      "/topics.json": "http://localhost:8080",
     },
   },
   css: {